Handle no_log values in Validator class
- Track no_log values in the Validator instance
- Sanitize error messages before returning
- move supporting functinos out of basic.py to avoid circular imports
